♻️ Refactor: SearchListItem 컴포넌트 분리 #4
